- Patch by naeg   - Added NPC Operation loot   - Added test data to test npc fighting   - Modified server loot handling to be able to work with superclient.

+++ src/server/bulkobjects/pscharacter.cpp
@@ -1675,7 +1675,7 @@ csArray<psItem*> psCharacter::RemoveLootItems(csArray<csString> categories)
 {
     csArray<psItem*> items;
 
-    for(size_t x = lootPending.GetSize()-1; x >= 0; x--)
+    for(int x = lootPending.GetSize()-1; x >= 0; x--)
     {
         if(!lootPending[x])
         {
